rename Property\Collection to Properties
